# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Wishaw YMCA Badge API
  version: 0.0.1
  description: |
    OpenAPI contract for the Wishaw YMCA Esports badging backend.

    This API is designed for a separate React frontend and uses session-based authentication.
    Successful login establishes a JSESSIONID cookie; subsequent calls must include that cookie.
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /api/v1/admin/groups:
    post:
      tags:
        - Groups
      summary: Create a group
      operationId: createGroup
      security:
        - SessionCookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGroupRequest'
      responses:
        '201':
          description: Group created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/ServerError'
    get:
      tags:
        - Groups
      summary: List groups
      operationId: listGroups
      security:
        - SessionCookieAuth: []
      responses:
        '200':
          description: Group list.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GroupResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/v1/admin/groups/{id}:
    put:
      tags:
        - Groups
      summary: Update a group
      operationId: updateGroup
      security:
        - SessionCookieAuth: []
      parameters:
        - $ref: '#/components/parameters/GroupId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGroupRequest'
      responses:
        '200':
          description: Updated group.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/v1/admin/modules:
    post:
      tags:
        - Modules
      summary: Create a module
      operationId: createModule
      security:
        - SessionCookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateModuleRequest'
      responses:
        '201':
          description: Module created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModuleSummaryResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
    get:
      tags:
        - Modules
      summary: List modules
      operationId: listModules
      security:
        - SessionCookieAuth: []
      responses:
        '200':
          description: Module list.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ModuleSummaryResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/v1/admin/modules/{id}:
    get:
      tags:
        - Modules
      summary: Get module detail
      operationId: getModule
      security:
        - SessionCookieAuth: []
      parameters:
        - $ref: '#/components/parameters/ModuleId'
      responses:
        '200':
          description: Module with challenges and schedule items.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModuleDetailResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
    put:
      tags:
        - Modules
      summary: Update a module
      operationId: updateModule
      security:
        - SessionCookieAuth: []
      parameters:
        - $ref: '#/components/parameters/ModuleId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateModuleRequest'
      responses:
        '200':
          description: Updated module.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModuleSummaryResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/v1/admin/modules/{id}/challenges:
    post:
      tags:
        - Modules
      summary: Add a challenge to a module
      operationId: createChallenge
      security:
        - SessionCookieAuth: []
      parameters:
        - $ref: '#/components/parameters/ModuleId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateChallengeRequest'
      responses:
        '201':
          description: Updated module detail including the new challenge.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModuleDetailResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/v1/admin/challenges/{id}:
    put:
      tags:
        - Modules
      summary: Update an existing challenge
      operationId: updateChallenge
      security:
        - SessionCookieAuth: []
      parameters:
        - $ref: '#/components/parameters/ChallengeId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateChallengeRequest'
      responses:
        '200':
          description: Updated module detail containing the challenge.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModuleDetailResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/v1/admin/modules/{id}/schedule-items:
    post:
      tags:
        - Modules
      summary: Add a schedule item to a module
      operationId: createScheduleItem
      security:
        - SessionCookieAuth: []
      parameters:
        - $ref: '#/components/parameters/ModuleId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateScheduleItemRequest'
      responses:
        '201':
          description: Updated module detail including the new schedule item.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModuleDetailResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/v1/admin/schedule-items/{id}:
    put:
      tags:
        - Modules
      summary: Update a schedule item
      operationId: updateScheduleItem
      security:
        - SessionCookieAuth: []
      parameters:
        - $ref: '#/components/parameters/ScheduleItemId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateScheduleItemRequest'
      responses:
        '200':
          description: Updated module detail containing the schedule item.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModuleDetailResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /api/v1/admin/import/csv/upload:
    post:
      tags:
        - Import
      summary: Upload a CSV file and create an import preview batch
      operationId: uploadCsv
      security:
        - SessionCookieAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Parsed batch preview.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportPreviewResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/v1/admin/import/{batchId}/map-players:
    post:
      tags:
        - Import
      summary: Map imported external player references to known player ids
      operationId: mapPlayers
      security:
        - SessionCookieAuth: []
      parameters:
        - $ref: '#/components/parameters/BatchId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImportPlayerMappingRequest'
      responses:
        '200':
          description: Updated import preview.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportPreviewResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/v1/admin/import/{batchId}/preview:
    get:
      tags:
        - Import
      summary: Get the current preview state of an import batch
      operationId: getImportPreview
      security:
        - SessionCookieAuth: []
      parameters:
        - $ref: '#/components/parameters/BatchId'
      responses:
        '200':
          description: Import preview.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportPreviewResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/v1/admin/import/{batchId}/commit:
    post:
      tags:
        - Import
      summary: Commit an import batch and create challenge awards
      operationId: commitImport
      security:
        - SessionCookieAuth: []
      parameters:
        - $ref: '#/components/parameters/BatchId'
      responses:
        '200':
          description: Import committed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportCommitResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/v1/admin/import/{batchId}/report:
    get:
      tags:
        - Import
      summary: Get the final import batch report
      operationId: getImportReport
      security:
        - SessionCookieAuth: []
      parameters:
        - $ref: '#/components/parameters/BatchId'
      responses:
        '200':
          description: Import report.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportPreviewResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /api/v1/leaderboards/group/{groupId}:
    get:
      tags:
        - Leaderboards
      summary: Get the leaderboard for one group
      operationId: getGroupLeaderboard
      security:
        - SessionCookieAuth: []
      parameters:
        - $ref: '#/components/parameters/GroupId'
      responses:
        '200':
          description: Group leaderboard.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LeaderboardEntryResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
components:
  securitySchemes:
    SessionCookieAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID
      description: Session cookie returned by POST /api/v1/auth/login.
  parameters:
...
    GroupId:
      name: groupId
      in: path
      required: true
      schema:
        type: integer
        format: int64
    ModuleId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
    ChallengeId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
    ScheduleItemId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
    BatchId:
      name: batchId
      in: path
      required: true
      schema:
        type: integer
        format: int64
  responses:
    BadRequest:
      description: Request was syntactically valid but semantically invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    ValidationError:
      description: Bean validation failed for one or more request fields.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Unauthorized:
      description: Authentication is missing, invalid, or expired.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Forbidden:
      description: Authenticated user does not have the required role.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    NotFound:
      description: The requested resource was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Conflict:
      description: Request conflicts with current state, usually a duplicate resource.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    ServerError:
      description: An unexpected backend error occurred.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
  schemas:
...
    CreateGroupRequest:
      type: object
      required:
        - name
        - centreId
      properties:
        name:
          type: string
        gameName:
          type: string
          nullable: true
        ageBand:
          type: string
          nullable: true
        centreId:
          type: integer
          format: int64
    CreateModuleRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        gameName:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
    CreateChallengeRequest:
      type: object
      required:
        - name
        - badgeCategoryId
        - points
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        badgeCategoryId:
          type: integer
          format: int64
        points:
          type: integer
          minimum: 0
        displayOrder:
          type: integer
          default: 0
    CreateScheduleItemRequest:
      type: object
      properties:
        weekNumber:
          type: integer
        sessionFocus:
          type: string
          nullable: true
        linkedChallengeId:
          type: integer
          format: int64
          nullable: true
        sessionPlanUrl:
          type: string
          nullable: true
        sessionSlidesUrl:
          type: string
          nullable: true
        displayOrder:
          type: integer
...
    ImportPlayerMappingRequest:
      type: object
      properties:
        playerMappings:
          type: object
          additionalProperties:
            type: integer
            format: int64
      example:
        playerMappings:
          player1: 1
          player2: 2
...
    GroupResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        gameName:
          type: string
          nullable: true
        ageBand:
          type: string
          nullable: true
        centreId:
          type: integer
          format: int64
        centreName:
          type: string
        active:
          type: boolean
    ModuleSummaryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        gameName:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        active:
          type: boolean
        approved:
          type: boolean
    ModuleDetailResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        gameName:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        active:
          type: boolean
        approved:
          type: boolean
        challenges:
          type: array
          items:
            $ref: '#/components/schemas/ModuleChallengeResponse'
        scheduleItems:
          type: array
          items:
            $ref: '#/components/schemas/ModuleScheduleItemResponse'
    ModuleChallengeResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
          nullable: true
        points:
          type: integer
        badgeCategoryId:
          type: integer
          format: int64
        badgeCategoryName:
          type: string
        displayOrder:
          type: integer
        active:
          type: boolean
    ModuleScheduleItemResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        weekNumber:
          type: integer
        sessionFocus:
          type: string
          nullable: true
        linkedChallengeId:
          type: integer
          format: int64
          nullable: true
        sessionPlanUrl:
          type: string
          nullable: true
        sessionSlidesUrl:
          type: string
          nullable: true
        displayOrder:
          type: integer
...
    LeaderboardEntryResponse:
      type: object
      properties:
        rank:
          type: integer
        playerId:
          type: integer
          format: int64
        displayName:
          type: string
        centreName:
          type: string
          nullable: true
        groupName:
          type: string
          nullable: true
        totalPoints:
          type: integer
        highestLevel:
          type: string
          nullable: true
    ImportPreviewResponse:
      type: object
      properties:
        batchId:
          type: integer
          format: int64
        fileName:
          type: string
        status:
          $ref: '#/components/schemas/ImportStatus'
        totalRows:
          type: integer
        validRows:
          type: integer
        errorRows:
          type: integer
        warningRows:
          type: integer
        unmappedPlayers:
          type: array
          items:
            type: string
        rows:
          type: array
          items:
            $ref: '#/components/schemas/ImportRowDetail'
    ImportRowDetail:
      type: object
      properties:
        rowNumber:
          type: integer
        status:
          $ref: '#/components/schemas/ImportRowStatus'
        message:
          type: string
        rawData:
          type: string
    ImportCommitResponse:
      type: object
      properties:
        batchId:
          type: integer
          format: int64
        status:
          $ref: '#/components/schemas/ImportStatus'
        totalAwardsCreated:
          type: integer
        totalPlayersAffected:
          type: integer
        message:
          type: string
    ApiErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string
...
    ImportStatus:
      type: string
      enum:
        - UPLOADED
        - PREVIEW_READY
        - COMMITTED
        - FAILED
    ImportRowStatus:
      type: string
      enum:
        - OK
        - WARNING
        - ERROR
        - SKIPPED